requirements.txt
kivy
python-docx
dropbox
